+++ ocaml-lsp-server/src/custom_requests/custom_request.ml
@@ -1,3 +1,4 @@
+module Construct = Req_construct
 module Hover_extended = Req_hover_extended
 module Infer_intf = Req_infer_intf
 module Merlin_call_compatible = Req_merlin_call_compatible
@@ -6,4 +7,5 @@ module Typed_holes = Req_typed_holes
 module Type_enclosing = Req_type_enclosing
 module Wrapping_ast_node = Req_wrapping_ast_node
 module Get_documentation = Req_get_documentation
+module Type_search = Req_type_search
 module Merlin_jump = Req_merlin_jump