+++ CHANGES.md
@@ -1,10 +1,18 @@
 # Unreleased
 
+- Add custom
+  [`ocamllsp/typeSearch`](/ocaml-lsp-server/docs/ocamllsp/typeSearch-spec.md) request (#1369)
+
+- Make MerlinJump code action configurable (#1376)
+
 - Add custom [`ocamllsp/merlinJump`](/ocaml-lsp-server/docs/ocamllsp/merlinJump-spec.md) request (#1374)
 
 ## Fixes
 
 - Fix fd leak in running external processes for preprocessing (#1349)
+- Fix prefix parsing for completion of object methods (#1363, fixes #1358)
+- Remove some duplicates in the `selectionRange` answers (#1368)
+
 
 # 1.19.0
 
@@ -15,10 +23,16 @@
 
 - Add support for OCaml 5.2 (#1233)
 
+- Add a code-action for syntactic and semantic movement shortcuts based on Merlin's Jump command (#1364)
+
 ## Fixes
 
 - Kill unnecessary ocamlformat processes with sigterm rather than sigint or
   sigkill (#1343)
+
+## Features
+
+- Add custom [`ocamllsp/construct`](https://github.com/ocaml/ocaml-lsp/blob/ocaml-lsp-server/docs/ocamllsp/construct-spec.md) request (#1348)
 
 # 1.18.0